# Faster primitive arrays