--- a/src/Nerdbank.MessagePack/Nerdbank.MessagePack.csproj
+++ b/src/Nerdbank.MessagePack/Nerdbank.MessagePack.csproj
@@ -14,6 +14,10 @@
     <PackageReference Include="System.IO.Pipelines" />
   </ItemGroup>
   <ItemGroup>
+    <None Update="Converters\ArraysOfPrimitivesConverters.Generated.tt">
+      <Generator>TextTemplatingFileGenerator</Generator>
+      <LastGenOutput>ArraysOfPrimitivesConverters.Generated.cs</LastGenOutput>
+    </None>
     <None Update="Converters\IntConverters.tt">
       <Generator>TextTemplatingFileGenerator</Generator>
       <LastGenOutput>IntConverters.cs</LastGenOutput>
@@ -46,6 +50,11 @@
     <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Update="Converters\ArraysOfPrimitivesConverters.Generated.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>ArraysOfPrimitivesConverters.Generated.tt</DependentUpon>
+    </Compile>
     <Compile Update="Converters\IntConverters.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>