… test/Benchmarks/ArraysOfPrimitives.cs
… test/Nerdbank.MessagePack.Tests/ArraysOfPrimitivesTests.cs