Optimize serialization of arrays of primitives

These are super-small values of predictable max size. The largest perf cost to these previously had been the very many `MessagePackWriter.Advance(int)` calls. But we can reduce that from one per array element to just one per array, which makes such arrays dramatically faster.

## Before

| Method           | Mean     | Error     | StdDev   | Gen0   | Allocated |
|----------------- |---------:|----------:|---------:|-------:|----------:|
| Bool_Deserialize | 37.32 μs |  8.097 μs | 0.444 μs | 1.1597 |   10024 B |
| Bool_Serialize   | 38.50 μs |  7.784 μs | 0.427 μs |      - |         - |
| Double_Serialize | 41.46 μs |  7.525 μs | 0.412 μs |      - |         - |
| Single_Serialize | 43.66 μs | 50.927 μs | 2.791 μs |      - |         - |
| Int16_Serialize  | 73.34 μs |  5.794 μs | 0.318 μs |      - |         - |
| Int32_Serialize  | 79.60 μs |  5.485 μs | 0.301 μs |      - |         - |
| Int64_Serialize  | 92.66 μs | 33.140 μs | 1.816 μs |      - |         - |
| Int8_Serialize   | 66.61 μs |  7.251 μs | 0.397 μs |      - |         - |
| UInt16_Serialize | 32.16 μs |  7.697 μs | 0.422 μs |      - |         - |
| UInt32_Serialize | 41.43 μs |  0.443 μs | 0.024 μs |      - |         - |
| UInt64_Serialize | 47.61 μs |  5.934 μs | 0.325 μs |      - |         - |

## After

| Method           | Mean     | Error     | StdDev   | Gen0   | Allocated |
|----------------- |---------:|----------:|---------:|-------:|----------:|
| Bool_Deserialize | 33.98 us |  2.365 us | 0.130 us | 1.1597 |   10024 B |
| Bool_Serialize   | 34.47 us | 15.759 us | 0.864 us |      - |         - |
| Double_Serialize | 28.51 us |  4.251 us | 0.233 us |      - |         - |
| Single_Serialize | 20.30 us |  1.865 us | 0.102 us |      - |         - |
| Int16_Serialize  | 39.02 us | 10.901 us | 0.598 us |      - |         - |
| Int32_Serialize  | 41.60 us |  6.761 us | 0.371 us |      - |         - |
| Int64_Serialize  | 48.85 us |  5.687 us | 0.312 us |      - |         - |
| Int8_Serialize   | 45.08 us |  3.616 us | 0.198 us |      - |         - |
| UInt16_Serialize | 19.83 us |  0.839 us | 0.046 us |      - |         - |
| UInt32_Serialize | 21.71 us |  1.565 us | 0.086 us |      - |         - |
| UInt64_Serialize | 29.96 us |  9.823 us | 0.538 us |      - |         - |